+++ doc/source/change-log.rst
@@ -5,6 +5,7 @@ Change Log
 
 Since Version 1.13
 ------------------
+* Add SIMD-oriented Fast Mersenne Twister (`SFMT <http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/SFMT/>`) generator.
 * Add complex normal (:meth:`~randomstate.prng.mt19937.complex_normal`)
 
 Version 1.13